ready for deployment

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import expressAsyncHandler from "express-async-handler";
 import Goal from "../models/goalModel.js";
+import User from "../models/userModel.js";
 
 // @ desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = expressAsyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json(goals);
 });
 
@@ -19,7 +20,8 @@ const setGoal = expressAsyncHandler(async (req, res) => {
         throw new Error('Please add a text field');
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     });
     res.status(200).json(goal);
 });
@@ -34,6 +36,21 @@ const updateGoal = expressAsyncHandler(async (req, res) => {
     {
         res.status(404);
         throw new Error('Goal not found');
+    }
+
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!req.user)
+    {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id)
+    {
+        res.status(401);
+        throw new Error('User not authorized');
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true });
     res.status(200).json(updatedGoal);
@@ -50,8 +67,22 @@ const deleteGoal = expressAsyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Goal not found');
     }
-    await Goal.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message: `Delete goal ${req.params.id}` });
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!req.user)
+    {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id)
+    {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
+    const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
+    res.status(200).json(deletedGoal);
 });
